requirements.txt
ansicolors==1.0.2
Babel==2.3.4
Beaker==1.8.1
bottle==0.12.13
cffi==1.4.2
configparser==3.5.0b2
cryptography==1.2.1
cssselect==0.9.1
Django==1.9.11
django-appconf==1.0.1
django-bootstrap-markdown==1.8.1
django-braces==1.8.1
django-cachalot==1.4.1
django-cors-headers==2.0.2
django-debug-toolbar==1.6
django-filter==0.10.0
django-graceful-session-cleanup==0.2.2
django-guardian==1.3
django-gulp==2.1.1
django-imagekit==3.2.6
django-jinja==2.1.2
django-markdown-deux==1.0.5
django-phonenumber-field==1.1.0
django-rest-framework-mongoengine==3.3.1
django-telegram-bot==0.6.0
django-thumbs==0.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
enum34==1.1.2
env-tools==2.0.0
funcsigs==0.4
future==0.15.2
futures==2.2.0
httpie==0.9.2
httplib2==0.10.3
idna==2.0
ipaddress==1.0.16
jdatetime==1.5
Jinja2==2.8
markdown2==2.3.0
MarkupSafe==0.23
mock==1.3.0
mongoengine==0.13.0
MySQL-python==1.2.5
nose==1.3.7
oauthlib==1.0.3
pbr==1.8.0
phonenumberslite==7.3.2
pilkit==1.1.12
Pillow==2.8.1
psutil==2.2.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
Pygments==2.0.2
PyJWT==1.4.2
pymongo==3.4.0
pyOpenSSL==0.15.1
pyparsing==2.0.3
python-dateutil==2.4.2
python-memcached==1.57
python-telegram-bot==3.2.0
pytz==2015.6
queuelib==1.4.2
requests==2.7.0
rncryptor==3.2.0
selenium==2.45.0
simplejson==3.10.0
six==1.9.0
sqlparse==0.2.2
tini==3.0.1
w3lib==1.13.0
